📝 chore: check if gettext is installed in build.sh

+++ build.sh
@@ -2,6 +2,12 @@
 
 set -e
 
+gettext --version
+if [ $? -ne 0 ]; then
+    echo gettext is required to build the documentation.
+    exit 127
+fi
+
 rm -rf docs/*
 cp index.html docs/
 cp CNAME docs/
